RENDEZVOUS-409: concurrent access management of slots
