+++ pom.xml
@@ -2,13 +2,12 @@
     <parent>
         <artifactId>lutece-global-pom</artifactId>
         <groupId>fr.paris.lutece.tools</groupId>
-        <version>4.0.4</version>
+        <version>4.0.6</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
     <groupId>fr.paris.lutece.plugins</groupId>
     <artifactId>plugin-appointment</artifactId>
-    <packaging>lutece-plugin</packaging>
-    <version>2.0.32-SNAPSHOT</version>
+    <version>2.1.0-SNAPSHOT</version>
     <name>Lutece appointment plugin</name>
     <properties>  
         <componentName>appointment</componentName>
