+++ src/java/fr/paris/lutece/plugins/appointment/business/slot/SlotDAO.java
@@ -54,6 +54,8 @@ public final class SlotDAO extends UtilDAO implements ISlotDAO
     private static final String SQL_QUERY_NEW_PK = "SELECT max(id_slot) FROM appointment_slot";
     private static final String SQL_QUERY_INSERT = "INSERT INTO appointment_slot (id_slot, starting_date_time, ending_date_time, is_open, is_specific, max_capacity, nb_remaining_places, nb_potential_remaining_places, nb_places_taken, id_form) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)";
     private static final String SQL_QUERY_UPDATE = "UPDATE appointment_slot SET starting_date_time = ?, ending_date_time = ?, is_open = ?, is_specific = ?, max_capacity = ?, nb_remaining_places = ?, nb_potential_remaining_places = ?, nb_places_taken = ?, id_form = ? WHERE id_slot = ?";
+    private static final String SQL_QUERY_UPDATE_POTENTIAL_REMAINING_PLACE = "UPDATE appointment_slot SET nb_potential_remaining_places = ? WHERE id_slot = ?";
+
     private static final String SQL_QUERY_DELETE = "DELETE FROM appointment_slot WHERE id_slot = ?";
     private static final String SQL_QUERY_SELECT_COLUMNS = "SELECT id_slot, starting_date_time, ending_date_time, is_open, is_specific, max_capacity, nb_remaining_places, nb_potential_remaining_places, nb_places_taken, id_form ";
     private static final String SQL_FROM_APPOINTMENT_SLOT = "FROM appointment_slot";
@@ -269,6 +271,28 @@ public final class SlotDAO extends UtilDAO implements ISlotDAO
             }
         }
         return slot;
+    }
+    
+    @Override
+    public void updatePotentialRemainingPlaces( int nbPotentialRemainingPlaces, int nIdSlot, Plugin plugin  ){
+    	
+    	DAOUtil daoUtil= null;
+    	try
+        {
+    		daoUtil = new DAOUtil( SQL_QUERY_UPDATE_POTENTIAL_REMAINING_PLACE, plugin );
+            daoUtil.setInt( 1, nbPotentialRemainingPlaces );
+            daoUtil.setInt( 2, nIdSlot );
+            daoUtil.executeUpdate( );
+           
+        }
+        finally
+        {
+           if( daoUtil != null ){
+            daoUtil.free();
+           }
+        }
+    	
+    	 
     }
 
     /**
