+++ src/java/fr/paris/lutece/plugins/appointment/resources/appointment_messages.properties
@@ -442,3 +442,6 @@ modifySlot.labelDate=Slot Date
 modifySlot.helpModifyEndSlot=If you change the ending time of the slot to a number that is not a multiple of the usual time, please select below the option you would like to use for the following slots.
 modifySlot.shiftSlot=Shift the following slots from the specified ending time.
 modifySlot.notShiftSlot=Do not shift the following slots.
+
+daemon.slotDaemon.name=Delete slot in memory Daemon
+daemon.slotDaemon.description=Delete slot in memory Daemon
