+++ src/java/fr/paris/lutece/plugins/appointment/resources/appointment_messages_fr.properties
@@ -470,3 +470,6 @@ modifySlot.helpModifyEndSlot=Si vous modifiez l'heure de fin du cr\u00e9neau \u0
                              cr\u00e9neaux suivants
 modifySlot.shiftSlot=D\u00e9caler les cr\u00e9neaux suivants \u00e0 partir de l'heure de fin renseign\u00e9e.
 modifySlot.notShiftSlot=Ne pas d\u00e9caler les cr\u00e9neaux suivants.
+
+daemon.slotDaemon.name=Deamon de suppression des slots dans la mémoire dont la date de début du créneau est inférieure a la date courante
+daemon.slotDaemon.description=Deamon de suppression des slots dans la mémoire dont la date de début du créneau est inférieure a la date courante
