+++ src/java/fr/paris/lutece/plugins/appointment/web/AppointmentSlotJspBean.java
@@ -66,6 +66,7 @@ import fr.paris.lutece.plugins.appointment.service.ClosingDayService;
 import fr.paris.lutece.plugins.appointment.service.DisplayService;
 import fr.paris.lutece.plugins.appointment.service.FormService;
 import fr.paris.lutece.plugins.appointment.service.ReservationRuleService;
+import fr.paris.lutece.plugins.appointment.service.SlotSafeService;
 import fr.paris.lutece.plugins.appointment.service.SlotService;
 import fr.paris.lutece.plugins.appointment.service.TimeSlotService;
 import fr.paris.lutece.plugins.appointment.service.WeekDefinitionService;
@@ -645,76 +646,83 @@ public class AppointmentSlotJspBean extends AbstractAppointmentFormAndSlotJspBea
     @Action( ACTION_DO_MODIFY_SLOT )
     public String doModifySlot( HttpServletRequest request )
     {
+        boolean bOpeningHasChanged = false;
         Slot slotFromSessionOrFromDb = null;
         String strIdSlot = request.getParameter( PARAMETER_ID_SLOT );
-        int nIdSlot = Integer.parseInt( strIdSlot );
-        if ( nIdSlot != 0 )
-        {
-            slotFromSessionOrFromDb = SlotService.findSlotById( nIdSlot );
-        }
-        else
-        {
-            slotFromSessionOrFromDb = (Slot) request.getSession( ).getAttribute( SESSION_ATTRIBUTE_SLOT );
-        }
         LocalTime endingTime = LocalTime.parse( request.getParameter( PARAMETER_ENDING_TIME ) );
         boolean bIsOpen = Boolean.parseBoolean( request.getParameter( PARAMETER_IS_OPEN ) );
         int nMaxCapacity = Integer.parseInt( request.getParameter( PARAMETER_MAX_CAPACITY ) );
         boolean bEndingTimeHasChanged = false;
-        boolean bOpeningHasChanged = false;
+    
         boolean bShiftSlot = Boolean.parseBoolean( request.getParameter( PARAMETER_SHIFT_SLOT ) );
-        if ( bIsOpen != slotFromSessionOrFromDb.getIsOpen( ) )
-        {
-            slotFromSessionOrFromDb.setIsOpen( bIsOpen );
-            bOpeningHasChanged = true;
-        }
-
-        // If we edit the slot, we need to check if this slot is not a closing
-        // day
-        ClosingDay closingDay = ClosingDayService.findClosingDayByIdFormAndDateOfClosingDay( slotFromSessionOrFromDb.getIdForm( ),
-                slotFromSessionOrFromDb.getDate( ) );
-        if ( closingDay != null )
-        {
-            // If the slot is a closing day, we need to remove it from the table
-            // closing day so that the slot is not in conflict with the
-            // definition of the closing days
-            ClosingDayService.removeClosingDay( closingDay );
-        }
-        if ( nMaxCapacity != slotFromSessionOrFromDb.getMaxCapacity( ) )
-        {
-            slotFromSessionOrFromDb.setMaxCapacity( nMaxCapacity );
-            // Need to set also the nb remaining places and the nb potential
-            // remaining places
-            // If the slot already exist, the good values will be set at the
-            // update of the slot with taking the old values
-            // If it is a new slot, the value set here will be good
-            slotFromSessionOrFromDb.setNbRemainingPlaces( nMaxCapacity );
-            slotFromSessionOrFromDb.setNbPotentialRemainingPlaces( nMaxCapacity );
-        }
-        LocalTime previousEndingTime = slotFromSessionOrFromDb.getEndingTime( );
-        if ( !endingTime.equals( previousEndingTime ) )
-        {
-            slotFromSessionOrFromDb.setEndingTime( endingTime );
-            slotFromSessionOrFromDb.setEndingDateTime( slotFromSessionOrFromDb.getDate( ).atTime( endingTime ) );
-            bEndingTimeHasChanged = true;
-        }
-        if ( bEndingTimeHasChanged && !checkNoAppointmentsOnThisSlotOrOnTheSlotsImpacted( slotFromSessionOrFromDb, bShiftSlot ) || bEndingTimeHasChanged
-                && !checkEndingTimeOfSlot( endingTime, slotFromSessionOrFromDb ) )
-        {
-            request.getSession( ).setAttribute( SESSION_ATTRIBUTE_SLOT, slotFromSessionOrFromDb );
-            return redirect( request, VIEW_MODIFY_SLOT, PARAMETER_ID_FORM, slotFromSessionOrFromDb.getIdForm( ) );
-        }
-        SlotService.updateSlot( slotFromSessionOrFromDb, bEndingTimeHasChanged, previousEndingTime, bShiftSlot );
-        AppLogService.info( LogUtilities.buildLog( ACTION_DO_MODIFY_SLOT, strIdSlot, getUser( ) ) );
-        addInfo( MESSAGE_INFO_SLOT_UPDATED, getLocale( ) );
-        boolean appointmentsImpacted = !AppointmentUtilities.checkNoValidatedAppointmentsOnThisSlot( slotFromSessionOrFromDb );
-        if ( appointmentsImpacted && bOpeningHasChanged )
-        {
-            addInfo( MESSAGE_INFO_VALIDATED_APPOINTMENTS_IMPACTED, getLocale( ) );
-        }
-        if ( appointmentsImpacted && nMaxCapacity < slotFromSessionOrFromDb.getNbPlacesTaken( ) )
-        {
-            addInfo( MESSAGE_INFO_SURBOOKING, getLocale( ) );
-        }
+        int nIdSlot = Integer.parseInt( strIdSlot );
+        Object lock = SlotSafeService.getLockOnSlot( nIdSlot );
+		synchronized (lock) {
+	        if ( nIdSlot != 0 )
+	        {
+	            slotFromSessionOrFromDb = SlotService.findSlotById( nIdSlot );
+	        }
+	        else
+	        {
+	            slotFromSessionOrFromDb = (Slot) request.getSession( ).getAttribute( SESSION_ATTRIBUTE_SLOT );
+	        }
+	       
+				
+	        if ( bIsOpen != slotFromSessionOrFromDb.getIsOpen( ) )
+	        {
+	            slotFromSessionOrFromDb.setIsOpen( bIsOpen );
+	            bOpeningHasChanged = true;
+	        }
+	
+	        // If we edit the slot, we need to check if this slot is not a closing
+	        // day
+	        ClosingDay closingDay = ClosingDayService.findClosingDayByIdFormAndDateOfClosingDay( slotFromSessionOrFromDb.getIdForm( ),
+	                slotFromSessionOrFromDb.getDate( ) );
+	        if ( closingDay != null )
+	        {
+	            // If the slot is a closing day, we need to remove it from the table
+	            // closing day so that the slot is not in conflict with the
+	            // definition of the closing days
+	            ClosingDayService.removeClosingDay( closingDay );
+	        }
+	        if ( nMaxCapacity != slotFromSessionOrFromDb.getMaxCapacity( ) )
+	        {
+	            slotFromSessionOrFromDb.setMaxCapacity( nMaxCapacity );
+	            // Need to set also the nb remaining places and the nb potential
+	            // remaining places
+	            // If the slot already exist, the good values will be set at the
+	            // update of the slot with taking the old values
+	            // If it is a new slot, the value set here will be good
+	            slotFromSessionOrFromDb.setNbRemainingPlaces( nMaxCapacity );
+	            slotFromSessionOrFromDb.setNbPotentialRemainingPlaces( nMaxCapacity );
+	        }
+	        LocalTime previousEndingTime = slotFromSessionOrFromDb.getEndingTime( );
+	        if ( !endingTime.equals( previousEndingTime ) )
+	        {
+	            slotFromSessionOrFromDb.setEndingTime( endingTime );
+	            slotFromSessionOrFromDb.setEndingDateTime( slotFromSessionOrFromDb.getDate( ).atTime( endingTime ) );
+	            bEndingTimeHasChanged = true;
+	        }
+	        if ( bEndingTimeHasChanged && !checkNoAppointmentsOnThisSlotOrOnTheSlotsImpacted( slotFromSessionOrFromDb, bShiftSlot ) || bEndingTimeHasChanged
+	                && !checkEndingTimeOfSlot( endingTime, slotFromSessionOrFromDb ) )
+	        {
+	            request.getSession( ).setAttribute( SESSION_ATTRIBUTE_SLOT, slotFromSessionOrFromDb );
+	            return redirect( request, VIEW_MODIFY_SLOT, PARAMETER_ID_FORM, slotFromSessionOrFromDb.getIdForm( ) );
+	        }
+	        SlotSafeService.updateSlot( slotFromSessionOrFromDb, bEndingTimeHasChanged, previousEndingTime, bShiftSlot );
+		 }
+	     AppLogService.info( LogUtilities.buildLog( ACTION_DO_MODIFY_SLOT, strIdSlot, getUser( ) ) );
+	     addInfo( MESSAGE_INFO_SLOT_UPDATED, getLocale( ) );
+	     boolean appointmentsImpacted = !AppointmentUtilities.checkNoValidatedAppointmentsOnThisSlot( slotFromSessionOrFromDb );
+	     if ( appointmentsImpacted && bOpeningHasChanged )
+	     {
+	    	 addInfo( MESSAGE_INFO_VALIDATED_APPOINTMENTS_IMPACTED, getLocale( ) );
+	     }
+	     if ( appointmentsImpacted && nMaxCapacity < slotFromSessionOrFromDb.getNbPlacesTaken( ) )
+	     {
+	         addInfo( MESSAGE_INFO_SURBOOKING, getLocale( ) );
+	     }
+	
         request.getSession( ).removeAttribute( SESSION_ATTRIBUTE_SLOT );
         Map<String, String> additionalParameters = new HashMap<>( );
         additionalParameters.put( PARAMETER_ID_FORM, Integer.toString( slotFromSessionOrFromDb.getIdForm( ) ) );
@@ -849,22 +857,25 @@ public class AppointmentSlotJspBean extends AbstractAppointmentFormAndSlotJspBea
 
         for ( Slot slotImpacted : listSlotsImpactedWithAppointments )
         {
-            // If the max capacity has changed,
-            // need to update it for all the slots that already have
-            // appointments
-            if ( bMaxCapacityHasChanged )
-            {
-                slotImpacted.setMaxCapacity( nMaxCapacity );
-                SlotService.updateRemainingPlaces( slotImpacted );
-            }
-            // if the opening of the timeslot has changed and there are
-            // appointments impacted,
-            // all the corresponding slots are marked as specific
-            if ( bOpeningHasChanged )
-            {
-                slotImpacted.setIsSpecific( bIsOpen );
-            }
-            SlotService.updateSlot( slotImpacted );
+        	Object lock = SlotSafeService.getLockOnSlot( slotImpacted.getIdSlot() );
+    		synchronized (lock) {
+	            // If the max capacity has changed,
+	            // need to update it for all the slots that already have
+	            // appointments
+	            if ( bMaxCapacityHasChanged )
+	            {
+	                slotImpacted.setMaxCapacity( nMaxCapacity );
+	                SlotSafeService.updateRemainingPlaces( slotImpacted );
+	            }
+	            // if the opening of the timeslot has changed and there are
+	            // appointments impacted,
+	            // all the corresponding slots are marked as specific
+	            if ( bOpeningHasChanged )
+	            {
+	                slotImpacted.setIsSpecific( bIsOpen );
+	            }
+	            SlotSafeService.updateSlot( slotImpacted );
+    	   }
         }
     }
