+++ webapp/WEB-INF/conf/plugins/appointment.properties
@@ -49,3 +49,7 @@ appointment.moduleAppointmentResource.name=appointment-resource
 appointment.calendarTemplates.calendarTemplatesFolder=skin/plugins/appointment/calendar/
 # Activer le workflow (workflow par défaut dans appointment)
 appointment.activate.workflow=false
+
+
+daemon.slotDaemon.interval=86400
+daemon.slotDaemon.onstartup=1
